requirements.txt
django==2.0.7
boto3
django-jsonfield
